+++ js/settings-admin.js
@@ -22,11 +22,29 @@
  *
  */
 
+function postSuccess(selector, id) {
+    $(selector).after(" <span id='" + id + "' class='msg success'>Saved</span>");
+    setTimeout(function(){
+        $("#" + id).remove()
+      }, 3000);
+}
+
+function postError(selector, id) {
+    $(selector).after(" <span id='" + id + "' class='msg error'>Error</span>");
+    setTimeout(function(){
+        $("#" + id).remove()
+      }, 3000);
+}
+
 window.addEventListener("DOMContentLoaded", function () {
     $("#breezedark-theme-enabled").change(function () {
         $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
             theme_enabled: this.checked ? 1 : 0,
             theme_login_page: $("#breezedark-theme-login-page").prop("checked") ? 1 : 0,
+        }).done(function() {
+            postSuccess("label[for='breezedark-theme-enabled']", "breezedark-theme-enabled-msg");
+        }).fail(function() {
+            postError("label[for='breezedark-theme-enabled']", "breezedark-theme-enabled-msg");
         });
 
         $("#breezedark-theme-login-page").prop("disabled", !$("#breezedark-theme-enabled").prop("checked"))
@@ -36,6 +54,20 @@ window.addEventListener("DOMContentLoaded", function () {
         $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
             theme_login_page: this.checked ? 1 : 0,
             theme_enabled: $("#breezedark-theme-enabled").prop("checked") ? 1 : 0,
+        }).done(function() {
+            postSuccess("label[for='breezedark-theme-login-page']", "breezedark-theme-login-page-msg");
+        }).fail(function() {
+            postError("label[for='breezedark-theme-login-page']", "breezedark-theme-login-page-msg");
+        });
+    });
+
+    $("#breezedark-theme-custom-styling-button").click(function () {
+        $.post(OC.generateUrl("apps/breezedark/settings/custom-styling"), {
+            theme_custom_styling: $("#breezedark-theme-custom-styling").val(),
+        }).done(function() {
+            postSuccess("#breezedark-theme-custom-styling-button", "breezedark-theme-custom-styling-button-msg");
+        }).fail(function() {
+            postError("#breezedark-theme-custom-styling-button", "breezedark-theme-custom-styling-button-msg");
         });
     });
 });